refactor(plugin): 12 个元数据访问器抽至 app.helper.plugin_metadata(S9c)
S9-decompose 第 2 切片(Metadata 簇)。把 god-object 中 12 个只读元数据聚合访问器(get_plugin_commands/apis/services/modules/actions/agent_tools/remote_entry/remotes/auth_providers/sidebar_nav/dashboard_meta/dashboard)抽成 plugin_metadata 模块级函数(首参 running_plugins)。

这些访问器对实例状态唯一耦合是 _running_plugins 一个字典;函数体逐字迁出(self._running_plugins → running_plugins,dict() 快照语义不变)。PluginManager 保留 12 个一行委托门面 → api/agent/command/scheduler/chain 等大量调用方 PluginManager().get_plugin_xxx() 零改动。PluginManager 零 monkey-patch(workflow 已证),抽取契约安全。

plugin_manager.py 1803→1444 行(god-object 累计 2026→1444)。

验证:py_compile;门面委托 + 12 函数就位探针;行为单测(fake 插件:聚合 + pid 注入 + get_state 过滤 + pid 过滤);新增 tests/test_plugin_metadata.py(7);dashboard/cloner/plugin/cycle/metainfo 105 passed。过程中 test_plugin_dashboard 捕获 Callable/Tuple 漏 import 并修正,印证行为覆盖有效。
